Rename `home-screen.tsx` to `HomeScreen.tsx`.
