+++ src/screens/home/index.ts
@@ -1,1 +1,1 @@
-export { default as HomeScreen } from "./home-screen";
+export { default as HomeScreen } from "./HomeScreen";
